+++ how-to/workspace-platform-starter/client/src/framework/modules.ts
@@ -6,6 +6,7 @@ import {
 } from "@openfin/workspace-platform";
 import { getApp, getApps } from "./apps";
 import { checkCondition, conditionChanged } from "./conditions";
+import * as Dialog from "./dialog";
 import { getEndpointClient } from "./endpoint";
 import type { EndpointClient } from "./endpoint-client";
 import * as favoriteProvider from "./favorite";
@@ -17,6 +18,7 @@ import * as Menu from "./menu";
 import { launchPage, launchView } from "./platform/browser";
 import type { PlatformApp, UpdatableLaunchPreference } from "./shapes/app-shapes";
 import type { ConditionContextTypes, ConditionsClient } from "./shapes/conditions-shapes";
+import type { DialogClient } from "./shapes/dialog-shapes";
 import type { FavoriteClient } from "./shapes/favorite-shapes";
 import type { Logger } from "./shapes/logger-shapes";
 import type { MenuClient } from "./shapes/menu-shapes";
@@ -31,7 +33,9 @@ import type {
 	ModuleTypes
 } from "./shapes/module-shapes";
 import type { NotificationClient } from "./shapes/notification-shapes";
+import type { ShareClient } from "./shapes/share-shapes";
 import type { ThemeClient } from "./shapes/theme-shapes";
+import * as Share from "./share";
 import {
 	getCurrentColorSchemeMode,
 	getCurrentIconFolder,
@@ -287,6 +291,8 @@ export function getDefaultHelpers(): ModuleHelpers {
 		getThemeClient,
 		getMenuClient,
 		getConditionsClient,
+		getShareClient,
+		getDialogClient,
 		launchApp: async (appId: string, launchPreference?: UpdatableLaunchPreference): Promise<void> => {
 			logger.info(`launchApp: Looking up appId: ${appId}`);
 			const app = await getApp(appId);
@@ -435,5 +441,29 @@ async function getConditionsClient(): Promise<ConditionsClient> {
 			return checkCondition(platform, conditionId, contextType);
 		},
 		changed: conditionChanged
+	};
+}
+
+/**
+ * Get the share client to use with the modules.
+ * @returns The share client.
+ */
+async function getShareClient(): Promise<ShareClient | undefined> {
+	if (Share.isShareEnabled()) {
+		return {
+			typeEnabled: Share.typeEnabled,
+			share: Share.share,
+			confirmation: Share.confirmation
+		};
+	}
+}
+
+/**
+ * Get the dialog client to use with the modules.
+ * @returns The dialog client.
+ */
+async function getDialogClient(): Promise<DialogClient> {
+	return {
+		showConfirmation: Dialog.showConfirmation
 	};
 }